+++ src/app/components/ExternalLink.css.ts
@@ -1,6 +1,7 @@
 import { createVar, style } from '@vanilla-extract/css'
 
 export const arrowColor = createVar('arrowColor')
+export const iconUrl = createVar('iconUrl')
 
 export const root = style({
   selectors: {
@@ -13,7 +14,7 @@ export const root = style({
       marginLeft: '0.325em',
       marginRight: '0.25em',
       width: '0.5em',
-      mask: 'url(/.vocs/icons/arrow-diagonal.svg) no-repeat center / contain',
+      mask: `${iconUrl} no-repeat center / contain`,
     },
   },
 })