+++ src/app/utils/hydrateLazyRoutes.ts
@@ -1,8 +1,8 @@
 import { type RouteObject, matchRoutes } from 'react-router-dom'
 
-export async function hydrateLazyRoutes(routes: RouteObject[]) {
+export async function hydrateLazyRoutes(routes: RouteObject[], basePath: string | undefined) {
   // Determine if any of the initial routes are lazy
-  const lazyMatches = matchRoutes(routes, window.location)?.filter((m) => m.route.lazy)
+  const lazyMatches = matchRoutes(routes, window.location, basePath)?.filter((m) => m.route.lazy)
 
   // Load the lazy matches and update the routes before creating your router
   // so we can hydrate the SSR-rendered content synchronously