+++ src/vite/preview.ts
@@ -12,11 +12,20 @@ type PreviewParameters = {
 
 export async function preview({ outDir = 'dist' }: PreviewParameters = {}) {
   const { config } = await resolveVocsConfig()
-  const { rootDir } = config
+  const { basePath, rootDir } = config
 
   const app = new Hono()
+
   app.use('*', compress())
-  app.use('/*', serveStatic({ root: resolve(rootDir, outDir) }))
+  app.use(
+    '/*',
+    serveStatic({
+      root: resolve(rootDir, outDir),
+      rewriteRequestPath(path) {
+        return basePath ? path.replace(basePath!, '') : path
+      },
+    }),
+  )
 
   return new Promise<ReturnType<typeof serve> & { port: number }>((res) => {
     async function createServer(port = 4173) {